+++ internal/products/app/command/add_product.go
@@ -14,7 +14,7 @@ type AddProduct struct {
 	title       string
 	description string
 	image       string
-	price       float64
+	price       float32
 	quantity    int
 }
