+++ internal/products/domain/product/repository.go
@@ -15,7 +15,7 @@ type Repository interface {
 		title string,
 		description string,
 		image string,
-		price float64,
+		price float32,
 		quantity int,
 	) error
 	UpdateProduct(
